[Nexus]: fix LUID structure data type issue.
Bug: b/378251125

(cherry picked from commit 516e9960a0d6b17d7cbfecf5f4c2c4144ea6576f)

+++ virtio_dxgkrnl/include/uapi/d3dkmthk.h
@@ -57,8 +57,8 @@ struct winluid {
 struct large_integer {
 	union {
 		struct {
-			unsigned long lowpart;
-			long highpart;
+			__u32 lowpart;
+			__s32 highpart;
 		} u;
 		int64_t quadpart;
 	};
